requirements.txt
flask==3.0.0
pinecone-client==3.0.0
openai==1.3.0
python-dotenv==1.0.0
tiktoken==0.5.2
PyPDF2==3.0.1 